More refactoring, add sidecar sweeper
